Check: comment status values in the database have flipped so they match node status
There is two places where the comments.status field is accessed.

The one in phpbb2drupal.module seems to use a value passed in via the
$edit function parameter, I presume from another part of Drupal, which
will be using the new logic, so no change needed.

The other is already Using symbolic constances as advised by upgrade
guide.

So there seems to be nothing to do.